requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
